fix lint errors and circular dependency between app-builder services
